@@ -6,7 +6,8 @@ class ItemConstants {
   static const String itemSubCategoryId = 'item_sub_category_id';
   static const String itemImages = 'item_images';
   static const String itemTitle = 'item_title';
-  static const String itemAddress = 'item_address';
+  static const String itemProvince = 'item_province';
+  static const String itemRegion = 'item_region';
   static const String itemTotalPrice = 'item_total_price';
   static const String itemPriceType = 'item_price_type';
   static const String itemSalePriceType = 'item_sale_price_type';
